combile lazy-loaded modules related to config editing
