+++ catalog/app/containers/Admin/Buckets/Tabulator/ConfigEditor.tsx
@@ -1,22 +1,22 @@
 import * as React from 'react'
+import * as FF from 'final-form'
 import * as RF from 'react-final-form'
 import * as M from '@material-ui/core'
 
-import { loadMode } from 'components/FileEditor/loader'
+import tabulatorTableSchema from 'schemas/tabulatorTable.yml.json'
 
-const TextEditor = React.lazy(() => import('components/FileEditor/TextEditor'))
+import TextEditor from 'components/FileEditor/TextEditor'
+import { loadMode } from 'components/FileEditor/loader'
+import { JsonInvalidAgainstSchema } from 'utils/error'
+import { makeSchemaValidator } from 'utils/json-schema'
+import * as yaml from 'utils/yaml'
 
 const TEXT_EDITOR_TYPE = { brace: 'yaml' as const }
 
 type ConfigEditorFieldProps = RF.FieldRenderProps<string> &
   M.TextFieldProps & { className: string }
 
-export default function ConfigEditor({
-  errors,
-  input,
-  meta,
-  ...props
-}: ConfigEditorFieldProps) {
+export function ConfigEditor({ errors, input, meta, ...props }: ConfigEditorFieldProps) {
   loadMode('yaml')
 
   const error = meta.error || meta.submitError
@@ -39,4 +39,20 @@ export default function ConfigEditor({
       type={TEXT_EDITOR_TYPE}
     />
   )
+}
+
+export const validateTable: FF.FieldValidator<string> = (inputStr?: string) => {
+  try {
+    const data = yaml.parse(inputStr)
+    const validator = makeSchemaValidator(tabulatorTableSchema)
+    const errors = validator(data)
+    if (errors.length) {
+      return new JsonInvalidAgainstSchema({ errors }).message
+    }
+    return undefined
+  } catch (error) {
+    // eslint-disable-next-line no-console
+    console.error(error)
+    return 'invalid'
+  }
 }
